drawtool code refactoring
using abstract classes
