+++ src/main/java/jdraw/figures/LineTool.java
@@ -1,38 +1,15 @@
 package jdraw.figures;
 
 import jdraw.framework.DrawContext;
-import jdraw.framework.DrawTool;
-import jdraw.framework.DrawView;
-
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.MouseEvent;
 
-public class LineTool implements DrawTool {
-
-	private static final String IMAGES = "/images/";
-
-	private DrawContext context;
-
-	private DrawView view;
-
+public class LineTool extends AbstractDrawTool {
 	private Line newLine = null;
 
-	private Point anchor = null;
-
 	public LineTool(DrawContext context) {
-		this.context = context;
-		this.view = context.getView();
-	}
-
-	@Override
-	public void deactivate() {
-		this.context.showStatusText("");
-	}
-
-	@Override
-	public void activate() {
-		this.context.showStatusText("Line Mode");
+		super(context);
 	}
 
 	@Override
@@ -59,11 +36,6 @@ public class LineTool implements DrawTool {
 		this.context.showStatusText("Line Mode");
 	}
 
-	@Override
-	public Cursor getCursor() {
-		return Cursor.getPredefinedCursor(Cursor.CROSSHAIR_CURSOR);
-	}
-	
 	@Override
 	public Icon getIcon() {
 		return new ImageIcon(getClass().getResource(IMAGES + "line.png"));
