+++ src/main/java/jdraw/figures/OvalTool.java
@@ -6,38 +6,15 @@
 package jdraw.figures;
 
 import jdraw.framework.DrawContext;
-import jdraw.framework.DrawTool;
-import jdraw.framework.DrawView;
-
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.MouseEvent;
 
-public class OvalTool implements DrawTool {
-
-	private static final String IMAGES = "/images/";
-
-	private DrawContext context;
-
-	private DrawView view;
-
+public class OvalTool extends AbstractDrawTool {
 	private Oval newOval = null;
 
-	private Point anchor = null;
-
 	public OvalTool(DrawContext context) {
-		this.context = context;
-		this.view = context.getView();
-	}
-
-	@Override
-	public void deactivate() {
-		this.context.showStatusText("");
-	}
-
-	@Override
-	public void activate() {
-		this.context.showStatusText("Oval Mode");
+		super(context);
 	}
 
 	@Override
@@ -65,11 +42,6 @@ public class OvalTool implements DrawTool {
 	}
 
 	@Override
-	public Cursor getCursor() {
-		return Cursor.getPredefinedCursor(Cursor.CROSSHAIR_CURSOR);
-	}
-	
-	@Override
 	public Icon getIcon() {
 		return new ImageIcon(getClass().getResource(IMAGES + "oval.png"));
 	}
@@ -78,5 +50,4 @@ public class OvalTool implements DrawTool {
 	public String getName() {
 		return "Oval";
 	}
-
 }
